requirements.txt
Flask==3.0.0
requests==2.31.0
opencv-python==4.9.0.80
moviepy==1.0.3
Pillow==10.1.0
SpeechRecognition==3.10.0
gTTS==2.4.0
pytesseract==0.3.10
numpy==1.26.2
gunicorn==21.2.0
